requirements.txt
pandas
django
djangorestframework
psycopg2-binary
sqlalchemy
python-dotenv
django-extensions
django-cors-headers